LayoutTests/media/video-src-remove-expected.txt

Test that removing src attribute triggers load
EVENT(emptied)
EVENT(loadstart)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
END OF TEST
